Extracted changelog options.
